sample revision

+++ Sample/ItemWork.cs
@@ -3,7 +3,6 @@ using Greatbone.Core;
 
 namespace Greatbone.Sample
 {
-    ///
     public abstract class ItemWork<V> : Work where V : ItemVarWork
     {
         protected ItemWork(WorkContext wc) : base(wc)
@@ -14,7 +13,7 @@ namespace Greatbone.Sample
 
 
     [Ui("货架")]
-    [User(User.OPRJOB)]
+    [User(User.OPRBASE)]
     public class OprItemWork : ItemWork<OprItemVarWork>
     {
         public OprItemWork(WorkContext wc) : base(wc)
@@ -23,14 +22,18 @@ namespace Greatbone.Sample
 
         public void @default(ActionContext ac)
         {
-            short shopid = ac[typeof(ShopVarWork)];
+            short shopid = ac[-1];
             using (var dc = ac.NewDbContext())
             {
-                const int proj = Item.SHOPID;
-                dc.Sql("SELECT ").columnlst(Item.Empty, proj)._("FROM items WHERE shopid = @1");
-                if (dc.Query(p => p.Set(shopid)))
+                if (dc.Query("SELECT * FROM items WHERE shopid = @1", p => p.Set(shopid)))
                 {
-                    ac.GiveGridPage(200, dc.ToArray<Item>(proj), (h, o) => { });
+                    ac.GiveGridPage(200, dc.ToArray<Item>(), (h, o) =>
+                    {
+                        h.COL("名称", o.name);
+                        h.COL("简述", o.descr);
+                        h.COL("价格", o.price);
+                        h.COL("成分", o.name);
+                    });
                 }
                 else
                 {
@@ -39,7 +42,7 @@ namespace Greatbone.Sample
             }
         }
 
-        [Ui("新建产品", Mode = UiMode.AnchorShow)]
+        [Ui("新建", Mode = UiMode.ButtonShow)]
         public async Task @new(ActionContext ac)
         {
             if (ac.GET)
@@ -48,44 +51,45 @@ namespace Greatbone.Sample
                 ac.GivePane(200, m =>
                 {
                     m.FORM_();
-                    m.TEXT(nameof(o.name), o.name, label: "品名", max: 10);
-                    m.TEXT(nameof(o.descr), o.descr, label: "描述", max: 30);
-                    m.TEXT(nameof(o.unit), o.unit, label: "单位（如：斤，小瓶）", required: true);
-                    m.NUMBER(nameof(o.price), o.price, label: "单价", required: true);
-                    m.NUMBER(nameof(o.min), o.min, label: "起订数量", min: (short) 1);
-                    m.NUMBER(nameof(o.step), o.step, label: "递增因子", min: (short) 1);
-                    m.NUMBER(nameof(o.max), o.max, label: "本批供应量");
-                    m.SELECT(nameof(o.status), o.status, Item.STATUS);
-
+                    m.TEXT(nameof(o.name), o.name, "名称", max: 10, required: true);
+                    m.TEXT(nameof(o.descr), o.descr, "简述", max: 30, required: true);
+                    m.TEXT(nameof(o.unit), o.unit, "单位", required: true);
+                    m.NUMBER(nameof(o.price), o.price, "单价", required: true);
+                    m.NUMBER(nameof(o.min), o.min, "起订数量", min: (short) 1);
+                    m.NUMBER(nameof(o.step), o.step, "增减间隔", min: (short) 1);
+                    m.NUMBER(nameof(o.max), o.max, "剩余供给");
+                    m.SELECT(nameof(o.status), o.status, Item.STATUS, "状态");
                     m._FORM();
                 });
             }
             else // post
             {
                 var o = await ac.ReadObjectAsync<Item>();
-                o.shopid = ac[typeof(ShopVarWork)];
+                o.shopid = ac[-1];
                 using (var dc = Service.NewDbContext())
                 {
-                    const int proj = Item.SHOPID;
-                    dc.Sql("INSERT INTO items")._(Item.Empty, proj)._VALUES_(Item.Empty, proj);
-                    dc.Execute(p => o.Write(p, proj));
-                    ac.GivePane(201);
+                    dc.Sql("INSERT INTO items")._(Item.Empty)._VALUES_(Item.Empty);
+                    dc.Execute(p => o.Write(p));
                 }
+                ac.GivePane(200); // close dialog
             }
         }
 
         [Ui("删除", Mode = UiMode.ButtonConfirm)]
         public async Task del(ActionContext ac)
         {
-            string shopid = ac[typeof(ShopVarWork)];
+            short shopid = ac[-1];
             var f = await ac.ReadAsync<Form>();
             string[] key = f[nameof(key)];
-            using (var dc = ac.NewDbContext())
+            if (key != null)
             {
-                dc.Sql("DELETE FROM items WHERE shopid = @1 AND name")._IN_(key);
-                dc.Execute(p => p.Set(shopid));
-                ac.GiveRedirect();
+                using (var dc = ac.NewDbContext())
+                {
+                    dc.Sql("DELETE FROM items WHERE shopid = @1 AND name")._IN_(key);
+                    dc.Execute(p => p.Set(shopid));
+                }
             }
+            ac.GiveRedirect();
         }
     }
 
@@ -102,7 +106,7 @@ namespace Greatbone.Sample
             short shopid = ac[typeof(ShopVarWork)];
             using (var dc = ac.NewDbContext())
             {
-                const int proj = Item.SHOPID;
+                const int proj = Item.UNMOD;
                 dc.Sql("SELECT ").columnlst(Item.Empty, proj)._("FROM items WHERE shopid = @1");
                 if (dc.Query(p => p.Set(shopid)))
                 {
